# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Spring System
  description: Spring System Sample Application.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:9966/system/api
...
paths:
...
  /specialties:
    get:
      tags:
        - specialty
      operationId: listSpecialties
      summary: Lists specialties
      description: Returns an array of specialty .
      responses:
        200:
          description: Specialties found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Specialty"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
        - specialty
      operationId: addSpecialty
      summary: Create a specialty
      description: Creates a specialty .
      requestBody:
        description: The specialty
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Specialty"
        required: true
      responses:
        200:
          description: Specialty created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Specialty not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /specialties/{specialtyId}:
    get:
      tags:
        - specialty
      operationId: getSpecialty
      summary: Get a specialty by ID
      description: Returns the specialty or a 404 error.
      parameters:
        - name: specialtyId
          in: path
          description: The ID of the speciality.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Specialty details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Specialty not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    put:
      tags:
        - specialty
      operationId: updateSpecialty
      summary: Update a specialty by ID
      description: Returns the specialty or a 404 error.
      parameters:
        - name: specialtyId
          in: path
          description: The ID of the specialty.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      requestBody:
        description: The pet
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Specialty"
        required: true
      responses:
        200:
          description: Specialty details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Specialty not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      tags:
        - specialty
      operationId: deleteSpecialty
      summary: Delete a specialty by ID
      description: Returns the specialty or a 404 error.
      parameters:
        - name: specialtyId
          in: path
          description: The ID of the specialty.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Specialty details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Specialty not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
    ProblemDetail:
      title: Problem Detail
      description: The schema for all error responses.
      type: object
      properties:
        type:
          title: Type
          description: Full URL that originated the error response.
          type: string
          format: uri
          example: "http://localhost:9966/system/api/owner"
          readOnly: true
        title:
          title: Title
          description: The short error title.
          type: string
          example: NoResourceFoundException
          readOnly: true
        status:
          title: Status
          description: HTTP status code
          type: integer
          format: int32
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 500
          readOnly: true
        detail:
          title: Detail
          description: The long error message.
          type: string
          example: "No static resource api/owner."
          readOnly: true
        timestamp:
          title: Timestamp
          description: The time the error occurred.
          type: string
          format: date-time
          example: "2024-11-23T13:59:21.3820407Z"
          readOnly: true
        schemaValidationErrors:
          title: Schema validation errors
          description: Validation errors against the OpenAPI schema.
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
      required:
        - type
        - title
        - status
        - detail
        - timestamp
        - schemaValidationErrors
    ValidationMessage:
      title: Validation message
      description: Messages describing a validation error.
      type: object
      properties:
        message:
          title: Message
          description: The validation message.
          type: string
          example: "[Path '/lastName'] Instance type (null) does not match any allowed primitive type (allowed: ['string'])"
          readOnly: true
      required:
        - message
      additionalProperties: true
    Specialty:
      title: Specialty
      description: Fields of specialty of vets.
      type: object
      properties:
        id:
          title: ID
          description: The ID of the specialty.
          type: integer
          format: int32
          minimum: 0
          example: 1
          readOnly: true
        name:
          title: Name
          description: The name of the specialty.
          type: string
          maxLength: 80
          minLength: 1
          example: radiology
      required:
        - id
        - name
...